MVCF: Fully switch to PatchSets
